[Refactor] Tab 모델 이전
