+++ AsyncSwift/Models/Tab.swift
@@ -1,0 +1,35 @@
+//
+//  Tab.swift
+//  AsyncSwift
+//
+//  Created by Inho Choi on 2022/10/08.
+//
+
+import Foundation
+
+enum Tab: String, CaseIterable {
+     case event = "Event"
+     case ticketing = "Ticketing"
+     case stamp = "Stamp"
+
+    var title: String {
+        rawValue
+    }
+
+    var systemImageName: String {
+        switch self {
+        case .event: return "calendar"
+        case .ticketing: return "banknote"
+        case .stamp: return "checkmark.square"
+        }
+    }
+
+    @ViewBuilder
+    var view: some View {
+        switch self {
+        case .event: EventView()
+        case .ticketing: TicketingView()
+        case .stamp: StampView()
+        }
+    }
+}
