+++ AsyncSwift/Observed/MainTabView+Observed.swift
@@ -86,30 +86,3 @@ extension MainTabView {
         }
     }
 }
-
-enum Tab: String, CaseIterable {
-     case event = "Event"
-     case ticketing = "Ticketing"
-     case stamp = "Stamp"
-
-    var title: String {
-        rawValue
-    }
-
-    var systemImageName: String {
-        switch self {
-        case .event: return "calendar"
-        case .ticketing: return "banknote"
-        case .stamp: return "checkmark.square"
-        }
-    }
-
-    @ViewBuilder
-    var view: some View {
-        switch self {
-        case .event: EventView()
-        case .ticketing: TicketingView()
-        case .stamp: StampView()
-        }
-    }
-}
